// distribuidora-web/src/app/core/api/generated/models/pos-product-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PosProductResponse {
  availableStock?: number;
  barcode?: string | null;
  brandId?: string;
  brandName?: string | null;
  categoryId?: string;
  categoryName?: string | null;
  favorite?: boolean;
  id?: string;
  minimumStock?: number;
  name?: string | null;
  price?: number;
  sku?: string | null;
  soldQuantity?: number;
}